feat: remove old DatePicker and let DatePicker2 take over
