--- frontend/src/components/DatePicker/Calendar/index.ts
@@ -1,4 +1,0 @@
-export * from './CalendarContext'
-export * from './CalendarHeader'
-export * from './CalendarPanel'
-export * from './CalendarTodayButton'
