+++ frontend/src/components/DatePicker/index.ts
@@ -1,1 +1,2 @@
-export { DateInput as default } from './DateInput'
+export * from './DatePicker'
+export type { DatePickerBaseProps } from './types'
