--- frontend/src/components/DatePicker2/index.ts
@@ -1,2 +1,0 @@
-export * from './DatePicker'
-export type { DatePickerBaseProps } from './types'
